refactor(browser): rename entrypoint to `@nuxt/test-utils/browser`
