+++ package.json
@@ -18,8 +18,8 @@
     "./module": "./src/module.ts",
     "./runtime": "./src/runtime-utils/index.ts",
     "./vitest-environment": "./src/vitest-environment.ts",
-    "./vitest-browser-nuxt": "./src/vitest-browser-nuxt/index.ts",
-    "./vitest-browser-nuxt/pure": "./src/vitest-browser-nuxt/pure.ts",
+    "./browser": "./src/browser/index.ts",
+    "./browser/pure": "./src/browser/pure.ts",
     "./package.json": "./package.json"
   },
   "imports": {
@@ -51,11 +51,11 @@
       "vitest-environment": [
         "./dist/vitest-environment.d.mts"
       ],
-      "vitest-browser-nuxt": [
-        "./dist/vitest-browser-nuxt/index.d.mts"
+      "browser": [
+        "./dist/browser/index.d.mts"
       ],
-      "vitest-browser-nuxt/pure": [
-        "./dist/vitest-browser-nuxt/pure.d.mts"
+      "browser/pure": [
+        "./dist/browser/pure.d.mts"
       ]
     }
   },
@@ -202,8 +202,8 @@
       "./module": "./dist/module.mjs",
       "./runtime": "./dist/runtime-utils/index.mjs",
       "./vitest-environment": "./dist/vitest-environment.mjs",
-      "./vitest-browser-nuxt": "./dist/vitest-browser-nuxt/index.mjs",
-      "./vitest-browser-nuxt/pure": "./dist/vitest-browser-nuxt/pure.mjs",
+      "./browser": "./dist/browser/index.mjs",
+      "./browser/pure": "./dist/browser/pure.mjs",
       "./package.json": "./package.json"
     }
   }
