+++ src/browser/index.ts
